move table column definition to separate file
